Import0xB826: Replace LERandomAccess with ByteBuffer

+++ src/test/java/it/smartphonecombo/uecapabilityparser/importer/nr/Import0xB826Test.kt
@@ -11,7 +11,7 @@ internal class Import0xB826Test {
     private fun parse(
         srcFilename: String,
         version: Int,
-        logSize: Long,
+        logSize: Int,
         numCombos: Int,
         totalCombos: Int?,
         index: Int?,
